+++ src/main/kotlin/com/perflog/domain/review/model/Review.kt
@@ -2,26 +2,28 @@ package com.perflog.domain.review.model
 
 import com.perflog.common.model.BaseTimeEntity
 import com.perflog.domain.member.model.Member
+import com.perflog.domain.perfume.model.entity.Perfume
 import jakarta.persistence.*
 
 @Entity
 @Table(name = "reviews")
 class Review(
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long = 0,
-
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id", nullable = false)
     var member: Member,
 
-    @Column(name = "perfume_id", nullable = false)
-    val perfumeId: Long = 0,
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "perfume_id", nullable = false)
+    val perfume: Perfume,
 
     @Column(nullable = false)
     var rating: Int,
 
     @Column(columnDefinition = "TEXT")
     var content: String
-) : BaseTimeEntity()
+) : BaseTimeEntity() {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long = 0
+}
